+++ client/app/app.js
@@ -36,6 +36,8 @@ import preload from './components/manager-preload/manager-preload.module';
 import dedicatedServerBandwidth from './dedicated/server/bandwidth/bandwidth.module';
 import dedicatedServerInterfaces from './dedicated/server/interfaces/interfaces.module';
 import dedicatedServerServers from './dedicated/server/servers/servers.module';
+import mfaEnrollment from './mfaEnrollment/mfaEnrollment.module';
+
 
 Environment.setRegion(__WEBPACK_REGION__);
 
@@ -61,6 +63,7 @@ angular
     'Module.ip',
     'Module.license',
     'Module.otrs',
+    mfaEnrollment,
     'ngCkeditor',
     'ngMessages',
     ngAtInternet,