+++ client/app/dedicated/server/servers/servers.controller.js
@@ -21,6 +21,14 @@ export default class ServersCtrl {
 
     this.stateEnumFilter = this.getEnumFilter(this.serverStateEnum, 'server_configuration_state_');
     this.datacenterEnumFilter = this.getEnumFilter(this.datacenterEnum, 'server_datacenter_');
+
+    this.columnsConfig = [
+      { name: 'name', sortable: this.getSorting('name') },
+      { name: 'reverse', sortable: this.getSorting('reverse') },
+      { name: 'commercialRange', sortable: this.getSorting('commercialRange') },
+      { name: 'datacenter', sortable: this.getSorting('datacenter') },
+      { name: 'state', sortable: this.getSorting('state') },
+    ];
   }
 
   static toUpperSnakeCase(str) {
@@ -41,13 +49,12 @@ export default class ServersCtrl {
   }
 
   getSorting(property) {
-    return this.sort === property ? this.sortOrder : '';
+    return this.sort === property ? this.sortOrder.toLowerCase() : '';
   }
 
   loadServers() {
     const currentOffset = this.paginationNumber * this.paginationSize;
     _.set(this.ouiDatagridService, 'datagrids.dg-servers.paging.offset', currentOffset < this.paginationTotalCount ? currentOffset : this.paginationTotalCount);
-
     return this.$q.resolve({
       data: _.get(this.dedicatedServers, 'data'),
       meta: {