+++ webpack.config.js
@@ -85,5 +85,8 @@ module.exports = (env = {}) => {
         __WEBPACK_REGION__: `'${env.region.toUpperCase()}'`,
       }),
     ],
+    optimization: {
+      runtimeChunk: 'single',
+    },
   });
 };